push

+++ core/src/com/game/Constants.java
@@ -1,5 +1,5 @@
 package com.game;
 
 public class Constants {
-	public static final int CAMERA_VIEW = 200;
+	public static final float CAMERA_VIEW = 200;
 }
